// src/Features/editableProfileCard/Model/consts/consts.ts
// for enums consts wcich compiles in object

export const enum  ValidateProfileError {
  INCORRECT_FIRST_NAME = "INCORRECT_FIRST_NAME",
  INCORRECT_LAST_NAME = "INCORRECT_LAST_NAME",
  INCORRECT_PATRONYMIC = "INCORRECT_PATRONYMIC",
  INCORRECT_DATE_BIRTHDAY="INCORRECT_DATE_BIRTHDAY",
  INCORRECT_ADDRESS="INCORRECT_ADDRESS",
  INCORRECT_PHONE_NUMBER="INCORRECT_PHONE_NUMBER",
  NO_DATA = "NO_DATA",
  SERVER_ERROR = "SERVER_ERROR",
}